fix: u128 amounts end-to-end, QUIC session recovery, oracle price consistency
Amount handling
- Widen the signing path to u128 on both platforms. JS passes atoms as a
  decimal string; iOS parses to (lo, hi) via `parseU128Halves`; Android
  JNI parses via `u128::from_str`. Eliminates the silent u64 wrap that
  turned 1000 SOV into 3.87 SOV on-chain.
- Centralize amount sanitization in `TokenService.coerceAmountAtoms` +
  `tokenUnits.validateAtomsString`. Every `amount: Number(...)` site in
  the send / mint / burn / wallet-transfer / DAO-stake / domain-register
  flows now routes through the single validator.
- Rebuild libquic_jni.so for arm64-v8a / armeabi-v7a / x86_64.

QUIC session recovery
- Retry loop in `quic.rawRequest` for authenticated requests that fail
  with framing errors or 401 "Invalid counter — possible replay":
  drops the UHP session via `cancelAll` and re-handshakes, up to 3
  attempts. Gracefully degrades on the wallet list / tx list / rewards
  paths that were locking the SID screen out on stale sessions.
- Android `cancelAll` now clears `quinnRequestQueue` +
  `quinnHandshakeInProgress` (iOS already did) so `resetAuthSession`
  actually forces a fresh handshake on the next call.
- `useAsyncData` catch binding renamed + normalized before the setState
  closure — fixes the Hermes "Property 'error' doesn't exist" crash on
  the SID tab when wallet fetches failed.

Activity & oracle price display
- `WalletTransaction` carries optional `decimals` / `token_id` / `symbol`
  so the activity list can render non-SOV tokens without hardcoding 18.
  `formatTxValue` takes decimals; `resolveTxDecimals` looks up per-tx
  decimals from the user's token balances when the backend omits them.
- New `formatAtomicPrice` / `formatAtomicPriceDisplay` in `tokenUnits.ts`:
  exact BigInt long-division of `(price_atomic, price_scale)` with a
  configurable `minFractionDigits` (default 8). No float conversion, no
  lazy digit trimming.
- Dashboard trending tokens + Oracle PriceTab both render via the same
  formatter on the same atomic-pair input — bit-identical strings for
  the same fetch.

Oracle dashboard UX
- New `useOracleData` hook: AsyncStorage-backed cache, auto-polling,
  graceful degradation. Keeps the last value on screen during refetch
  and when the node is unreachable — no more placeholder flicker.
- New `RefreshRing` atom: subtle SVG ring that fills toward the next
  auto-refetch, pulses while fetching, turns amber when stale, tap to
  retry. Replaces the full-screen spinner on every tab.

+++ android/app/src/main/rust/quic-jni/src/identity_bridge.rs
@@ -171,7 +171,7 @@ pub fn build_token_create_transaction(
     identity_json: &str,
     name: &str,
     symbol: &str,
-    initial_supply: u64,
+    initial_supply: u128,
     decimals: u8,
     treasury_recipient: [u8; 32],
     chain_id: u8,
@@ -181,7 +181,7 @@ pub fn build_token_create_transaction(
         &identity,
         name,
         symbol,
-        initial_supply as u128,
+        initial_supply,
         decimals,
         treasury_recipient,
         chain_id,
@@ -194,7 +194,7 @@ pub fn build_token_mint_transaction(
     identity_json: &str,
     token_id: &[u8],
     to_pubkey: &[u8],
-    amount: u64,
+    amount: u128,
     chain_id: u8,
 ) -> Result<String> {
     let identity = deserialize_identity(identity_json)?;
@@ -213,7 +213,7 @@ pub fn build_token_mint_transaction(
         to_arr[..to_pubkey.len()].copy_from_slice(to_pubkey);
     }
 
-    zhtp_client::build_mint_tx(&identity, &token_id_arr, &to_arr, amount as u128, chain_id)
+    zhtp_client::build_mint_tx(&identity, &token_id_arr, &to_arr, amount, chain_id)
         .map_err(|e| anyhow::anyhow!("Failed to build token mint transaction: {}", e))
 }
 
@@ -222,7 +222,7 @@ pub fn build_token_transfer_transaction(
     identity_json: &str,
     token_id: &[u8],
     to_pubkey: &[u8],
-    amount: u64,
+    amount: u128,
     chain_id: u8,
     nonce: u64,
 ) -> Result<String> {
@@ -242,7 +242,7 @@ pub fn build_token_transfer_transaction(
         to_arr[..to_pubkey.len()].copy_from_slice(to_pubkey);
     }
 
-    zhtp_client::build_transfer_tx(&identity, &token_id_arr, &to_arr, amount as u128, chain_id, nonce)
+    zhtp_client::build_transfer_tx(&identity, &token_id_arr, &to_arr, amount, chain_id, nonce)
         .map_err(|e| anyhow::anyhow!("Failed to build token transfer transaction: {}", e))
 }
 
@@ -250,7 +250,7 @@ pub fn build_token_transfer_transaction(
 pub fn build_token_burn_transaction(
     identity_json: &str,
     token_id: &[u8],
-    amount: u64,
+    amount: u128,
     chain_id: u8,
 ) -> Result<String> {
     let identity = deserialize_identity(identity_json)?;
@@ -262,6 +262,6 @@ pub fn build_token_burn_transaction(
         token_id_arr[..token_id.len()].copy_from_slice(token_id);
     }
 
-    zhtp_client::build_burn_tx(&identity, &token_id_arr, amount as u128, chain_id)
+    zhtp_client::build_burn_tx(&identity, &token_id_arr, amount, chain_id)
         .map_err(|e| anyhow::anyhow!("Failed to build token burn transaction: {}", e))
 }
